refactor: change amount and initialBalance attribute type to fit with new migration change in database

+++ src/schemas/account.schema.ts
@@ -7,7 +7,7 @@ export const accountSchema = z.object({
   type: z.enum(['MERCHANT', 'COMMON'], {
     message: 'must be MERCHANT or COMMON',
   }),
-  initialBalance: z.number().optional(),
+  initialBalance: z.number().int().optional(),
   createdAt: z.date(),
 });
